requirements.txt
anyio==4.8.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
colorama==0.4.6
deep-translator==1.11.4
Django==5.1.5
django-ckeditor-5==0.2.15
django-js-asset==3.0.1
django-redis==5.4.0
djangorestframework==3.15.2
gunicorn==20.1.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.7
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pytest==8.3.4
redis==5.2.1
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
